requirements.txt
# requirements.txt
opencv-python>=4.5
Pillow>=9.0
PyYAML>=6.0
ultralytics>=8.0
transformers>=4.30
torch>=1.13
huggingface-hub>=0.15
easyocr>=1.5
jinja2>=3.0
selenium>=4.0
chromedriver-binary>=114.0
webcolors
beautifulsoup4
